move helper model classes to proper location
