+++ wannadb_ui/common.py
@@ -300,120 +300,105 @@ class InformationPopup(QMainWindow):
         self.resize(1000, 700)
 
 
-class BestMatchUpdate:
-    def __init__(self, old_best_match: str, new_best_match: str, count: int):
-        self._old_best_match: str = old_best_match
-        self._new_best_match: str = new_best_match
-        self._count: int = count
+class InfoDialog(QDialog):
+    def __init__(self):
+        super().__init__()
 
-    @property
-    def old_best_match(self) -> str:
-        return self._old_best_match
+        self.dialog_shown: bool = False
 
-    @property
-    def new_best_match(self) -> str:
-        return self._new_best_match
+        self.info_list = None
+        self.image_list = None
+        self.current_index = 0
 
-    @property
-    def count(self) -> int:
-        return self._count
+        # Set up the dialog layout
+        self.layout = QVBoxLayout()
 
+        # Set a fixed width for the dialog
+        self.setFixedWidth(400)  # Set the fixed width you prefer
 
-class ThresholdPositionUpdate:
-    def __init__(self, best_guess: str,
-                 old_position: Optional[ThresholdPosition], new_position: ThresholdPosition,
-                 old_distance: Optional[float], new_distance: float,
-                 count: int):
-        self._best_guess: str = best_guess
-        self._old_position: Optional[ThresholdPosition] = old_position
-        self._new_position: ThresholdPosition = new_position
-        self._old_distance: float = old_distance
-        self._new_distance: float = new_distance
-        self._count: int = count
+        # Label to display the information text
+        self.info_label = QLabel()
+        self.info_label.setWordWrap(True)  # Enable word wrap for the label
+        self.layout.addWidget(self.info_label)
 
-    @property
-    def best_guess(self) -> str:
-        return self._best_guess
+        # Widget to display the PNG image
+        self.image_widget = QLabel()
+        self.layout.addWidget(self.image_widget)
 
-    @property
-    def old_position(self) -> Optional[ThresholdPosition]:
-        return self._old_position
+        # Buttons for navigation (Previous, Next, Skip)
+        self.button_layout = QHBoxLayout()
 
-    @property
-    def new_position(self) -> ThresholdPosition:
-        return self._new_position
+        self.prev_button = QPushButton("Previous")
+        self.prev_button.clicked.connect(self.show_previous)
+        self.button_layout.addWidget(self.prev_button)
 
-    @property
-    def old_distance(self) -> Optional[float]:
-        return self._old_distance
+        self.next_button = QPushButton("Next")
+        self.next_button.clicked.connect(self.show_next)
+        self.button_layout.addWidget(self.next_button)
 
-    @property
-    def new_distance(self) -> float:
-        return self._new_distance
+        self.skip_button = QPushButton("Skip")
+        self.skip_button.clicked.connect(self.skip)
+        self.button_layout.addWidget(self.skip_button)
 
-    @property
-    def count(self) -> int:
-        return self._count
+        # Add button layout to the main layout
+        self.layout.addLayout(self.button_layout)
 
+        # Set the layout for the dialog
+        self.setLayout(self.layout)
 
-class NewlyAddedNuggetContext:
-    def __init__(self, nugget: InformationNugget,
-                 old_distance: Union[float, None],
-                 new_distance: float,
-                 added_reason: AddedReason):
-        self._nugget = nugget
-        self._old_distance = old_distance
-        self._new_distance = new_distance
-        self._added_reason = added_reason
+    # Setter method to set the info_list
+    def set_info_list(self, info_list):
+        self.info_list = info_list
+        self.update_info()
 
-    @property
-    def nugget(self):
-        return self._nugget
+    # Setter method to set the image_list
+    def set_image_list(self, image_list):
+        self.image_list = image_list
+        self.update_image()
 
-    @property
-    def old_distance(self):
-        return self._old_distance
+    # Method to update the displayed information
+    def update_info(self):
+        if self.info_list is not None:
+            self.info_label.setText(self.info_list[self.current_index])
+            self.update_image()
+        self.update_buttons()
 
-    @property
-    def new_distance(self):
-        return self._new_distance
+    # Method to update the displayed PNG image
+    def update_image(self):
+        if self.image_list is not None:
+            image_path = self.image_list[self.current_index]
+            if image_path and image_path.endswith(".png"):
+                pixmap = QPixmap(image_path)
+                self.image_widget.setPixmap(pixmap)
+                self.image_widget.setVisible(True)
+            else:
+                self.image_widget.clear()
+                self.image_widget.setVisible(False)
 
-    @property
-    def added_reason(self):
-        return self._added_reason
+    # Method to update the state of the buttons
+    def update_buttons(self):
+        if self.info_list is not None:
+            self.prev_button.setEnabled(self.current_index > 0)
+            self.next_button.setEnabled(self.current_index < len(self.info_list) - 1)
 
+    # Method to show the previous piece of information
+    def show_previous(self):
+        if self.current_index > 0:
+            self.current_index -= 1
+            self.update_info()
 
-class NuggetUpdatesContext:
-    def __init__(self,
-                 newly_added_nugget_contexts: List[NewlyAddedNuggetContext],
-                 best_match_updates: List[BestMatchUpdate],
-                 threshold_position_updates: List[ThresholdPositionUpdate]):
-        self._newly_added_nugget_contexts: List[NewlyAddedNuggetContext] = newly_added_nugget_contexts
-        self._best_match_updates: List[BestMatchUpdate] = best_match_updates
-        self._threshold_position_updates: List[ThresholdPositionUpdate] = threshold_position_updates
+    # Method to show the next piece of information
+    def show_next(self):
+        if self.current_index < len(self.info_list) - 1:
+            self.current_index += 1
+            self.update_info()
 
-    @property
-    def newly_added_nugget_contexts(self) -> List[NewlyAddedNuggetContext]:
-        return self._newly_added_nugget_contexts
+    # Method to skip and close the dialog
+    def skip(self):
+        self.accept()
 
-    @property
-    def best_match_updates(self) -> List[BestMatchUpdate]:
-        return self._best_match_updates
-
-    @property
-    def threshold_position_updates(self) -> List[ThresholdPositionUpdate]:
-        return self._threshold_position_updates
-
-
-class AccessibleColor:
-    def __init__(self, color: QColor, corresponding_accessible_color: QColor):
-        self._color = color
-        self._corresponding_accessible_color = corresponding_accessible_color
-
-    @property
-    def color(self):
-        return self._color
-
-    @property
-    def corresponding_accessible_color(self):
-        return self._corresponding_accessible_color
+    # Override exec to prevent multiple executions
+    def exec(self):
+        if not self.dialog_shown:
+            super().exec()
+            self.dialog_shown = True
